--- junit-parameters/src/main/java/com/lezhnin/junit/parameters/SetSupplier.java
@@ -1,20 +1,0 @@
-package com.lezhnin.junit.parameters;
-
-import static java.util.stream.Collectors.toSet;
-import static java.util.stream.Stream.generate;
-import java.lang.reflect.Array;
-import java.util.function.Supplier;
-
-public class SetSupplier extends BaseSupplier {
-
-    SetSupplier(final Supplier<?> supplier, final Class<?> parameterType, final int maxSize) {
-        super(supplier, parameterType, maxSize);
-    }
-
-    @Override
-    public Object get() {
-        return generate(getSupplier())
-                .limit(getRandom().nextInt(getMaxSize()))
-                .collect(toSet());
-    }
-}
