+++ junit-parameters/src/main/java/com/lezhnin/junit/parameters/factory/SetSupplier.java
@@ -1,12 +1,12 @@
-package com.lezhnin.junit.parameters;
+package com.lezhnin.junit.parameters.factory;
 
-import static java.util.stream.Collectors.toList;
+import static java.util.stream.Collectors.toSet;
 import static java.util.stream.Stream.generate;
 import java.util.function.Supplier;
 
-public class ListSupplier extends BaseSupplier {
+public class SetSupplier extends BaseSupplier {
 
-    ListSupplier(final Supplier<?> supplier, final Class<?> parameterType, final int maxSize) {
+    SetSupplier(final Supplier<?> supplier, final Class<?> parameterType, final int maxSize) {
         super(supplier, parameterType, maxSize);
     }
 
@@ -14,6 +14,6 @@ public class ListSupplier extends BaseSupplier {
     public Object get() {
         return generate(getSupplier())
                 .limit(getRandom().nextInt(getMaxSize()))
-                .collect(toList());
+                .collect(toSet());
     }
 }
